[debops.users] Fix user not put into its primary group if user__acountrs[].group is omitted
According to documentation:

user__acounts[].group

Optional. Name of the UNIX group which will be set as the primary group of
a given account. If group is not specified, name will be used automatically
to create the corresponding UNIX group.

However, when group was omitted, name was used to create the corresponding
UNIX group but is was not added to it.